Added draft inheritance
First lines of values are automatically trimmed, so the TRIMLIST key is no
longer necessary (although not yet removed).

@INCLUDEFILE= now specifies a file to be included at that location in the
hash.  The file is used to provide default keys for the current location.
Once the current location finishes being defined, the file is then merged
in with it.  This creates a rudimentary and somewhat simple form of
class based inheritance.

@KEY += now adds the value to the pre-existing key, or generates the key if it
doesn't exist yet.

Within values, other values may be referenced via:
@$.KEY : references a key in the current locality.  Could also reference
	@$THIS.KEY, but @$.KEY is simpler
@$+KEY : references a key in the inheritance files
@$/KEY : references a key at the top level of the design
@$KEY  : first looks for the key here, then in any inherited maps, and then
works its way up the tree looking for the key.  The first value found is the
value used.