Checking in working copies
Moved global data into globals.
Moved the clock info into its own file.
Still working on building hte bus info capability.
Drafted an initial subbus capability.  A subbus is a peripheral with both
  slave and master interfaces.  A classic example would be a bus translator
  that changes things like the bus width, the bus clock, or even the bus
  type.  This capability is still far from complete.
I now have new address assignment code in plist.  This works well for a
list of peripherals, and it keeps the number of LUT's necessary to decode
a peripheral to a minimum.  Still to figure out is how to create a
single or double clock bus.